Extract WHO_AM_I listener
