+++ src/background/background.ts
@@ -40,7 +40,6 @@ import initDeploymentUpdater from "@/background/deploymentUpdater";
 import initTheme from "@/background/initTheme";
 import initStarterMods from "@/background/starterMods";
 import { initPartnerTokenRefresh } from "@/background/partnerIntegrations";
-import { initContentScriptRawMessagesListener } from "@/background/contentScript";
 import { initLogSweep } from "@/telemetry/logging";
 import { initModUpdater } from "@/background/modUpdater";
 import { initRuntimeLogging } from "@/development/runtimeLogging";
@@ -68,7 +67,6 @@ void initNavigation();
 initExecutor();
 initFeatureFlagBackgroundListeners();
 initContextMenus();
-initContentScriptRawMessagesListener();
 initBrowserCommands();
 initDeploymentUpdater();
 initTheme();
